Try harder to lazily require 'gpgme'
If the 'gpgme' gem is being installed via puppet, then the gem might not
be available at autoload time. If `require 'gpgme'` fails at autoload
time the failure is outputted as warning, and the require is deferred
till prefetching. If the game still isn't available then an exception
will be raised as normal. This way the gpg module can bootstrap the
provider.