Update agent to more accurately locate a running agent
This commit updates the gpg agent class in order to more accurately
determine whether an agent is running by using the gpg-agent command.
Without this change ps is used to look for gpg-agent, which can be
difficult to locate because jobs run that consume the gpg agent. When
this happens the agent does not start because it incorrectly determines
that it is already running.
Additionally, the agent manifest is re-arragned to make more clear the
commands that are running and the defaults for Exec are removed in favor
of specifying the parameters within the exec declarations themselves in
order to ensure that those parameters do not apply unintentionally to
other execs in the catalog.