Add management of the gpg-agent package
This commit adds management of the gpg-agent package for debian and ubuntu. Additionally, a flag is provided in order to enable or disable package management of the agent. Additionally, spacing is added in the agent manifest in order to make it more readable. Without this change the agent package must be deployed manually in order for gpg::agent to work.